refactor: Function-specific file's primary functions should match filename.
