+++ app/src/main/java/no/usn/mob3000/ui/components/info/ContentDisplay.kt
@@ -1,0 +1,81 @@
+package no.usn.mob3000.ui.components.info
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+/**
+ * Content display for news, documentation and faq.
+ *
+ * @author 258030
+ * @created 2024-10-30
+ */
+@Composable
+fun ContentDisplay(
+    modifier: Modifier = Modifier,
+    title: String,
+    summary: String?,
+    content: String?,
+    createdAt: Long? = null,
+    modifiedAt: Long? = null
+) {
+    Box(
+        modifier = modifier
+            .fillMaxSize()
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp)
+                .verticalScroll(rememberScrollState())
+        ) {
+            Text(
+                text = title,
+                fontSize = 24.sp,
+                fontWeight = FontWeight.Bold,
+                modifier = Modifier.padding(bottom = 16.dp)
+            )
+
+            summary?.let {
+                Text(
+                    text = it,
+                    fontSize = 16.sp,
+                    fontWeight = FontWeight.Medium,
+                    fontStyle = FontStyle.Italic,
+                    modifier = Modifier.padding(bottom = 16.dp)
+                )
+            }
+
+            content?.let {
+                Text(
+                    text = it,
+                    fontSize = 14.sp,
+                    modifier = Modifier.padding(bottom = 16.dp)
+                )
+            }
+
+            createdAt?.let {
+                Text(
+                    text = formatInstant(it, isCreated = true),
+                    fontSize = 12.sp,
+                    modifier = Modifier.padding(bottom = 4.dp)
+                )
+            }
+
+            modifiedAt?.let {
+                Text(
+                    text = formatInstant(it, isCreated = false),
+                    fontSize = 12.sp,
+                    modifier = Modifier.padding(bottom = 4.dp)
+                )
+            }
+        }
+    }
+}
